@@ -1,6 +1,6 @@
 package com.notmarra.notcredits.events;
 
-import com.notmarra.notcredits.Notcredits;
+import com.notmarra.notcredits.NotCredits;
 import com.notmarra.notcredits.util.DatabaseManager;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
@@ -11,12 +11,12 @@ import org.bukkit.event.player.PlayerJoinEvent;
 public class PlayerJoin implements Listener {
     @EventHandler
     public void onPlayerJoin(PlayerJoinEvent event) {
-        Bukkit.getScheduler().runTaskAsynchronously(Notcredits.getInstance(), new Runnable() {
+        Bukkit.getScheduler().runTaskAsynchronously(NotCredits.getInstance(), new Runnable() {
             @Override
             public void run() {
                 Player player = event.getPlayer();
-                if (!DatabaseManager.getInstance(Notcredits.getInstance()).hasAccount(player.getUniqueId().toString())){
-                    DatabaseManager.getInstance(Notcredits.getInstance()).setupPlayer(player.getUniqueId().toString(), player.getName(), Notcredits.getInstance().getConfig().getDouble("default_balance"));
+                if (!DatabaseManager.getInstance(NotCredits.getInstance()).hasAccount(player.getUniqueId().toString())){
+                    DatabaseManager.getInstance(NotCredits.getInstance()).setupPlayer(player.getUniqueId().toString(), player.getName(), NotCredits.getInstance().getConfig().getDouble("default_balance"));
                 }
             }
         });
